+++ killport.js
@@ -1,14 +1,17 @@
 #!/usr/bin/env node
 
 const net = require('net')
+const path = require('path')
 const { spawnSync } = require('child_process')
+const pkg = require(path.join(__dirname, 'package.json'))
 
 // parse command-line arguments
 let silent = false
 for (const arg of process.argv) {
-  if (arg === '--silent') {
-    silent = true
-    break
+  if (arg === '--silent') silent = true
+  if (arg === '-v' || arg === '-version' || arg === '--version') {
+    console.log(pkg.version)
+    process.exit(0)
   }
 }
 
@@ -21,15 +24,23 @@ if (isNaN(port)) {
 // function to check if a port is in use
 function checkPort (port, callback) {
   const server = net.createServer()
+  let ipv4Check
 
   server.once('error', (err) => {
     if (err.code === 'EADDRINUSE') callback(null, true) // port is in use
     else callback(err)
   })
 
-  server.once('listening', () => {
-    server.close()
-    callback(null, false) // port is not in use
+  server.on('listening', () => {
+    if (ipv4Check) {
+      server.close()
+      callback(null, false) // port is not in use
+    } else {
+      server.close(() => {
+        ipv4Check = true
+        server.listen(port, '0.0.0.0') // listen on ipv4 addresses
+      })
+    }
   })
 
   server.listen(port)
@@ -62,42 +73,39 @@ function findPidOnPort (port) {
     return null
   }
 
-  let pid
-  if (platform === 'win32') {
-    // extract PID on windows
-    const lines = result.stdout.trim().split('\n')
-    for (const line of lines) {
-      const parts = line.split(/\s+/)
-      if (parts[2] === `0.0.0.0:${port}` && parts[4] === 'LISTENING') {
-        pid = parseInt(parts[5])
-        break
-      }
+  const pids = []
+  const lines = result.stdout.trim().split('\n')
+  for (const line of lines) {
+    let pid
+    const parts = line.trim().split(/\s+/)
+    if (platform === 'win32') {
+      // extract PID on windows
+      if (parts[1] === `0.0.0.0:${port}` && parts[3] === 'LISTENING') pid = parseInt(parts[4])
+    } else {
+      // extract PID on *nix systems
+      pid = parseInt(parts[1])
     }
-  } else {
-    // extract PID on *nix systems
-    const lines = result.stdout.trim().split('\n')
-    const parts = lines[lines.length - 1].trim().split(/\s+/)
-    pid = parts[1]
+    if (pid) pids.push(pid)
   }
 
-  return pid
+  return pids
 }
 
 // function to kill the process running on the given port
 function killProcessOnPort (port) {
-  const pid = parseInt(findPidOnPort(port))
-  if (pid) {
-    try {
-      process.kill(pid, 'SIGKILL')
-      if (!silent) console.log(`Killed process ${pid} running on port ${port}`)
-    } catch (err) {
-      if (!silent) {
-        const error = err.code === 'EPERM' ? 'Permission denied' : err
-        console.error(`Error killing process ${pid}:`, error)
+  const pids = findPidOnPort(port)
+  if (pids) {
+    for (const pid of pids) {
+      try {
+        process.kill(pid, 'SIGKILL')
+        if (!silent) console.log(`Killed process ${pid} running on port ${port}`)
+      } catch (err) {
+        if (!silent) {
+          const error = err.code === 'EPERM' ? 'Permission denied' : err
+          console.error(`Error killing process ${pid}:`, error)
+        }
       }
     }
-  } else {
-    if (!silent) console.log(`No process found running on port ${port}`)
   }
 }